# docs/openapi.json
{
  "components": {
    "schemas": {
      "Assignment": {
        "properties": {
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "default": "practice",
            "title": "Mode",
            "type": "string"
          },
          "question_set_id": {
            "format": "uuid",
            "title": "Question Set Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentStatus",
            "default": "assigned"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds"
          }
        },
        "required": [
          "family_id",
          "question_set_id",
          "child_id"
        ],
        "title": "Assignment",
        "type": "object"
      },
      "AssignmentStatus": {
        "enum": [
          "assigned",
          "in_progress",
          "submitted",
          "grading",
          "results_ready",
          "correcting",
          "completed",
          "withdrawn",
          "stopped"
        ],
        "title": "AssignmentStatus",
        "type": "string"
      },
...
      "CreateAssignmentRequest": {
        "properties": {
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "mode": {
            "default": "practice",
            "enum": [
              "practice",
              "exam"
            ],
            "title": "Mode",
            "type": "string"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "maximum": 14400.0,
                "minimum": 60.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds"
          }
        },
        "required": [
          "child_id"
        ],
        "title": "CreateAssignmentRequest",
        "type": "object"
      },
...
      "CreateImportRequest": {
        "properties": {
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "filenames": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Filenames",
            "type": "array"
          },
          "purpose": {
            "$ref": "#/components/schemas/ImportPurpose"
          },
          "source_paths": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Source Paths",
            "type": "array"
          },
          "subject": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "filenames",
          "purpose",
          "title",
          "subject"
        ],
        "title": "CreateImportRequest",
        "type": "object"
      },
...
      "Family": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Family",
        "type": "object"
      },
      "FamilyInvitation": {
        "properties": {
          "accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "external_notification_sent": {
            "const": false,
            "default": false,
            "title": "External Notification Sent",
            "type": "boolean"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invited_by": {
            "title": "Invited By",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "family_id",
          "email",
          "invited_by"
        ],
        "title": "FamilyInvitation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportPurpose": {
        "enum": [
          "use_as_questions",
          "generate_similar"
        ],
        "title": "ImportPurpose",
        "type": "string"
      },
...
      "Question": {
        "properties": {
          "answer_key": {
            "additionalProperties": true,
            "title": "Answer Key",
            "type": "object"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "points": {
            "default": 1,
            "title": "Points",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "question_set_id": {
            "format": "uuid",
            "title": "Question Set Id",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          }
        },
        "required": [
          "family_id",
          "question_set_id",
          "position",
          "type",
          "prompt",
          "answer_key"
        ],
        "title": "Question",
        "type": "object"
      },
...
      "QuestionSet": {
        "properties": {
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionSetStatus"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "title",
          "subject",
          "status"
        ],
        "title": "QuestionSet",
        "type": "object"
      },
      "QuestionSetDraft": {
        "properties": {
          "question_set": {
            "$ref": "#/components/schemas/QuestionSet"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "question_set",
          "questions"
        ],
        "title": "QuestionSetDraft",
        "type": "object"
      },
      "QuestionSetImport": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "filenames": {
            "items": {
              "type": "string"
            },
            "title": "Filenames",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/ImportPurpose"
          },
          "question_set_id": {
            "format": "uuid",
            "title": "Question Set Id",
            "type": "string"
          },
          "source_paths": {
            "items": {
              "type": "string"
            },
            "title": "Source Paths",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionSetStatus",
            "default": "needs_review"
          }
        },
        "required": [
          "family_id",
          "question_set_id",
          "filenames",
          "purpose"
        ],
        "title": "QuestionSetImport",
        "type": "object"
      },
      "QuestionSetStatus": {
        "enum": [
          "draft",
          "processing",
          "needs_review",
          "confirmed"
        ],
        "title": "QuestionSetStatus",
        "type": "string"
      },
      "QuestionType": {
        "enum": [
          "single_choice",
          "multiple_choice",
          "typed_text",
          "word_order",
          "handwriting",
          "photo",
          "listening"
        ],
        "title": "QuestionType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Learning Assessment API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/invitations/pending": {
      "get": {
        "operationId": "list_pending_invitations_v1_invitations_pending_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FamilyInvitation"
                  },
                  "title": "Response List Pending Invitations V1 Invitations Pending Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pending Invitations",
        "tags": [
          "invitations"
        ]
      }
    },
    "/v1/invitations/{invitation_id}/accept": {
      "post": {
        "operationId": "accept_invitation_v1_invitations__invitation_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Family"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/v1/question-sets/imports": {
      "post": {
        "operationId": "create_import_v1_question_sets_imports_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionSetImport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Import",
        "tags": [
          "question-sets"
        ]
      }
    },
    "/v1/question-sets/{question_set_id}": {
      "get": {
        "operationId": "get_question_set_v1_question_sets__question_set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Set Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionSetDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Question Set",
        "tags": [
          "question-sets"
        ]
      }
    },
    "/v1/question-sets/{question_set_id}/assignments": {
      "post": {
        "operationId": "assign_question_set_v1_question_sets__question_set_id__assignments_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Set Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Question Set",
        "tags": [
          "question-sets"
        ]
      }
    },
    "/v1/question-sets/{question_set_id}/confirm": {
      "post": {
        "operationId": "confirm_question_set_v1_question_sets__question_set_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Set Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Question Set",
        "tags": [
          "question-sets"
        ]
      }
    },
...
  }
}